Complete image change and reorganize files
